+++ packages/leancode_debug_page/lib/src/core/logging_http_client.dart
@@ -8,8 +8,8 @@ import 'package:rxdart/rxdart.dart';
 class LoggingHttpClient extends http.BaseClient
     implements LogGatherer<RequestLogRecord> {
   LoggingHttpClient({http.Client? client})
-    : _httpClient = client ?? http.Client(),
-      _logsController = BehaviorSubject.seeded([]);
+      : _httpClient = client ?? http.Client(),
+        _logsController = BehaviorSubject.seeded([]);
 
   final http.Client _httpClient;
   final BehaviorSubject<List<RequestLogRecord>> _logsController;
@@ -37,19 +37,20 @@ class LoggingHttpClient extends http.BaseClient
     final responseBodyCompleter = Completer<String>();
 
     _logsController.add(
-      logs..add(
-        RequestLogRecord(
-          method: request.method,
-          url: request.url,
-          startTime: startTime,
-          endTime: endTime,
-          statusCode: response.statusCode,
-          requestHeaders: request.headers,
-          requestBody: requestBody,
-          responseBodyCompleter: responseBodyCompleter,
-          responseHeaders: response.headers,
+      logs
+        ..add(
+          RequestLogRecord(
+            method: request.method,
+            url: request.url,
+            startTime: startTime,
+            endTime: endTime,
+            statusCode: response.statusCode,
+            requestHeaders: request.headers,
+            requestBody: requestBody,
+            responseBodyCompleter: responseBodyCompleter,
+            responseHeaders: response.headers,
+          ),
         ),
-      ),
     );
 
     return http.StreamedResponse(