+++ packages/leancode_debug_page/lib/src/models/request_log_record.dart
@@ -43,7 +43,7 @@ class RequestLogRecord {
       (final code) when 200 <= code && code < 300 => RequestStatus.success,
       (final code) when 300 <= code && code < 400 => RequestStatus.redirect,
       (final code) when 400 <= code && code < 500 => RequestStatus.clientError,
-      (final code) when 500 <= code && code < 600 => RequestStatus.clientError,
+      (final code) when 500 <= code && code < 600 => RequestStatus.serverError,
       _ => RequestStatus.unknown
     };
   }