+++ packages/leancode_debug_page/lib/src/ui/logs_inspector/requests/request_details_screen/request_details_screen_response_tab.dart
@@ -21,7 +21,7 @@ class RequestDetailsScreenResponseTab extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return SingleChildScrollView(
-      child: DefaultTextStyle(
+      child: DefaultTextStyle.merge(
         style: DebugPageTypography.medium,
         child: Column(
           crossAxisAlignment: CrossAxisAlignment.start,