+++ packages/leancode_debug_page/lib/src/ui/typography.dart
@@ -3,10 +3,14 @@ import 'package:flutter/material.dart';
 class DebugPageTypography {
   static TextStyle _getTextStyle(double fontSize) => TextStyle(
         fontSize: fontSize,
-        color: Colors.black87,
       );
 
   static TextStyle get small => _getTextStyle(12);
   static TextStyle get medium => _getTextStyle(16);
   static TextStyle get large => _getTextStyle(20);
+
+  static TextStyle get codeBlock => _getTextStyle(12).copyWith(
+        fontFamily: 'monospace',
+        fontFamilyFallback: ['Monaco', 'Courier'],
+      );
 }