+++ packages/leancode_lint/lib/lints/avoid_single_child_in_multi_child_widget.dart
@@ -41,7 +41,7 @@ class _Visitor extends SimpleAstVisitor<void> {
   @override
   void visitInstanceCreationExpression(InstanceCreationExpression node) {
     final constructorName = node.constructorName.type;
-    if (constructorName.element2 case final typeElement?) {
+    if (constructorName.element case final typeElement?) {
       // is it something we want to complain about?
       final match = _complain.firstWhereOrNull(
         (e) => e.$2.isExactly(typeElement),
@@ -52,7 +52,7 @@ class _Visitor extends SimpleAstVisitor<void> {
 
       // does it have a children argument?
       var children = node.argumentList.arguments.firstWhereOrNull(
-        (e) => e.correspondingParameter?.name3 == match.$1,
+        (e) => e.correspondingParameter?.name == match.$1,
       );
       if (children == null) {
         return;