+++ packages/leancode_lint/lib/lints/use_design_system_item.dart
@@ -27,6 +27,9 @@
 //     required Iterable<ForbiddenItem> replacements,
 //   }) : super(
 //          lintCodeName: '${ruleName}_$preferredItemName',
+//          problemMessage: '{0} is forbidden within this design system.',
+//          correctionMessage:
+//              'Use the alternative defined in the design system: {1}.',
 //          replacements: {preferredItemName: replacements.toList()},
 //        );
 //
@@ -45,15 +48,5 @@
 //         replacements: entry.value,
 //       ),
 //     );
-//   }
-//
-//   @override
-//   String correctionMessage(String preferredItemName) {
-//     return 'Use the alternative defined in the design system: $preferredItemName.';
-//   }
-//
-//   @override
-//   String problemMessage(String itemName) {
-//     return '$itemName is forbidden within this design system.';
 //   }
 // }